Old version of Dockerfile for php container used alpine branch 3.10, which only has php 7.3 packages.  This fails when doing a composer update, as some packages require php 7.4+.
Updated to the alpine edge branch to get access to php 7.4.13 packages.  Also, fixed the names of a couple of packages (php7-mcrypt -> php7-pecl-mcrypt, etc).

Also in this push is a change to the composer.json - setting the php-json extension to >=1.7 doesn't work as expected.  It fails with my local environment.  It's reading my php-json as 7.4 and failing because it thinks it's not >= 1.7.  I changed this to an asterisk instead.